+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Journal des modifications
 - Affichage du nom complet d'un individu : choc de simplification en supprimant 2 options d'affichage dans le formatteur.
 - Rapport activité : augmentation à 10 min du timeout CSRF dans le formulaire de création/modification.
 - Possibilité de déclarer des missions d'enseignement
+- Les avis de soutenance deviennent des FichierThese afin de pouvoir être afficher sur la page des fichiers divers
 
 6.0.2
 -----